Rename cmdline source and add kcs_analyze
Renamed cmdline.py to KCSCmdline.py to follow the file naming
of all source code in the module.

Added a kcs_analyze program that allows to specify relevant
options (like parity and basefreq) on the command line and
dumps the result found in hex. This is useful to inspect yet
unknown protocols.