Add support for memory register data of FX502P
This changes the text and binary formats for saved file. We
need a file header. The header is 'FPnnn' for program data
and 'F nnn' for memory register data. This is consistent with
the actual headers when using the 3-digit number for saving
programs/data and displayed on the FX502P when loading them.