Add a bandpass filter to the device reader
It seems that the FX-502P can create some high pitched
noise, especially when the batteries get low. This leads
to immediately following sign change bits which confuses
the decoder. Adding a 1000Hz..2600Hz bandpass filter to
the rec(1) command fixes that.

TODO: Switch to using sox(1) instead for the wave module
for sound files. This allows to apply the same filter
to pre-recorded sound files.